requirements.txt
astroid==2.3.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pbr==5.4.1
pylint==2.4.4
requests==2.23.0
six==1.12.0
soupsieve==2.0
stevedore==1.30.1
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==16.7.2
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wrapt==1.11.2
lxml==4.5.0
google-cloud-firestoreastroid==2.3.3
beautifulsoup4==4.8.2
bs4==0.0.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
googleapis-common-protos==1.51.0
grpcio==1.27.2
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
pbr==5.4.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
pymongo==3.10.1
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.12.0
soupsieve==2.0
stevedore==1.30.1
TBB==0.1
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==16.7.2
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wrapt==1.11.2
